requirements.txt
biopython==1.81
matplotlib==3.7.1